backend/requirements.txt
cohere==5.5.3
qdrant-client==1.9.1
requests==2.31.0
beautifulsoup4==4.12.2
uvicorn==0.23.2
fastapi==0.103.1
pytest==7.4.2
python-dotenv==1.0.0
openai>=1.0.0
